Updated openstack/openstack Project: openstack/neutron  c4587ef4d1fb10bde9bc65e0bca87f66cd2fe5ed
Enable QoS scenario tests

This test has been disabled since [1].
But it seems the failure ratio is not so high with
the current code.  Let's re-enable it now and see how it goes.

Note: This same test has been enabled on the gate jobs for
other implementation for a while. It doesn't fail much there
either. (networking-midonet)

[1] I37f1488db258f6a4d383fb472cb5433c65371ac5

Related-Bug: #1662109
Change-Id: Ia39c73189ad8a3331c1911989fe69428f064f7a6