Updated openstack/openstack Project: openstack/nova  f17feb558c5a5d180e386733a8a4dd760bc68e99
Deprecate cells v1

With multi-cell support in Pike, we should deprecate cells v1
so we can at least start the deprecation signaling in the
docs and release notes. We may not end up removing cells v1
code in Queens, but this at least gives us the option.

Note that we also want to do this because nova-network cannot
be removed until we remove cells v1.

Change-Id: I1a173f7ce0715e684850e030c358e8175fa8724c