Updated openstack/openstack Project: openstack-infra/subunit2sql  4cbd7e4bc42d5e5f22d18abffc36616172ad4bbc
Add a unit test with a simple sql injection ala xkcd

This commit adds a basic unit tests with a simple sql injection attack
for a test name. All of the database interactions on the insert are done
through sqlalchemy and the ORM should be protecting us against this
attack vector. The test added here is just to sanity check this and so
we have something to point to. It's not intended to actually be a
thorough or even a real security test or a thorough security audit.

Change-Id: Iee5ed994328cf44834a1becb246f9983881d2129