Updated openstack/openstack Project: openstack/manila  98255e67152cb2d80f803d74cc78b82e109fe8a4
Removed unnecessary setUp() calls in tests


Basically, if you decide to override a method, that is because you want
to specify(or modify) the behavior of a method. If you just call the
inherited method from the overridden method, that is unnecessary.

Furthermore, every time when self.setUp() is called, now the super's
setUp() is called, which is basically exact the same behavior as
we have now.

TrivialFix

Change-Id: I9f4e03b93ff075c66b4e52342711c37f226d2f81