Updated openstack/openstack Project: openstack-infra/system-config  4edde02abfcae5521a1000b4a5664d0b1e07f637
Use new OpenStack logo on Gerrit sites

This puts the new logo in place for our gerrit deployments. Had to move
the header nav bar slightly to the right to get this to look well
without overlapping significantly. Note that this has been tested
locally using firefox to render the changes.

Change-Id: I1d773e137dd41ac2d7ee9347941d95732dae0932