Updated openstack/openstack Project: openstack/keystone  6b145664dd372e25972f8e774431e73deb7cdc00
Add cli/ directory for documentation

The specification detailing the great documentation migration
requires a doc/source/cli/ directory be added to the project.
The directory is intended for cli documentation.

This commit add the directory so that we are in compliance with the
specification but doesn't move any content into the directory. It is
instead treated like a placeholder until we can migrate the right
content into the cli guide.

Change-Id: Ida5ffc979d3c596547e9664dce8933d2cf9a4486
Closes-Bug: 1694460