Updated openstack/openstack Project: openstack/nova  2decfc77d8d0b2d71cf335344e8970b40d083c33
placement: remove existing allocs when set allocs

Bug #1707669 highlighted a situation that arose when attempting to
remove part of an allocation for a source host during a resize
operation where the exiting allocation was not being properly deleted.

In this patch, we remove the part of the WHERE condition that limited
deleted allocation records to only those referring to a particular
resource provider. In doing so, we make the creation of an allocation
for a consumer a proper overwrite operation.

Change-Id: I0835e5b4f22277465012aab9a5bf474608cb533b
Fixes-bug: #1707669