Updated openstack/openstack Project: openstack-infra/puppet-log_processor  1505f2e3b1766fdd74dc0b1f13846e2571ea9a0d
Add MQTT support to the gearman worker

This commit adds support to the gearman worker for publishing an mqtt
message when processing a gearman job succeeds or fails. It also adds
a message for when the processor passes the logs to logstash either via
stdout or over a socket. By default this is disabled since it requires
extra configuration to tell the worker how to talk to the mqtt broker.

Depends-On: Id0308d2d4d1843fcca73f459cffa2ae944bebd0c
Change-Id: I43be3562780c61591ebede61f3a8929e8217f199