Updated openstack/openstack Project: openstack-infra/system-config  694540c1030d55b5c05ec72f13a9dd004da848c9
Parameterize puppet version in install script

This patch adds a global PUPPET_VERSION variable to allow changing the
installed puppet version. The default version remains 3. CentOS and
Ubuntu support setting it to 4. Fedora already installs 4 by default
from the fedora repositories, and we defer worrying about opensuse and
gentoo for now since we do not use those operating systems in the
control plane and we no longer use puppet to build CI images.

The script adds the new "Puppet Collections" repository and changes the
name of the package to 'puppet-agent' when installing version 4. See the
install guide[1] for more details.

[1] https://docs.puppet.com/puppet/4.0/install_linux.html

Co-authored-by: Caleb Boylan <calebboylan@gmail.com>

Change-Id: Id829922875d495bca9ba11a3fe3428a7ebb2cd58