Updated openstack/openstack Project: openstack/ironic  87fe7cce4e39dff0596f7e74c299a7d296e1c651
Optimize node locking on heartbeat

Previously we uncoditionally tried to grab an exclusive lock on a node
on each heartbeat independent on whether any other action would have to
be performed.

This patch optimizes that so the lock is upgraded only when node is in
a provision state that might require some further action
(currently those are DEPLOYWAIT and CLEANWAIT).

Also, handling of deprecated 'agent_last_heartbeat' field
in driver_internal_info is removed.

Change-Id: If18b871aed63a9e5b41ccb9828ef8fa428f5ab48