Updated openstack/openstack Project: openstack/glance  e7ceaebd49a178e3657b24864a6924d6c294699d
doc: Explicitly set 'builders' option

A recent release of pbr explicitly requires stating which builders are
requested, rather than defaulting to html and man. This has resulted in
man pages not being built. Correct this oversight, along with a few
issues that have been introduced to 'doc/source/conf.py' since. It is
also necessary to remove a couple of plugins as there seems to be
another pbr bug under the hood. Fortunately none of these are used so
it's all good.

Change-Id: I22fd4d113704949dbf7704f50fa1d08fa10f11b6