Updated openstack/openstack Project: openstack/nova  a40c7cb6bce122b7a86d8e38d7d90e4aeef3f664
Test resize to same host with placement api

Test cases for resize to same host with confirm and revert scenarios
have been added.

The accounting of the current code is very incorrect at the moment, so this
asserts the existing (wrong) behavior, but marks the actual expected
values everywhere so that we can follow up with fixes to remove all of the
annotations.

Related-Bug: #1707252
Change-Id: Ica83ce5f43c0e7b84fd4117852e81e4d2979dea3