Updated openstack/openstack Project: openstack/horizon  53dd2dbd39c50b665ebe2d2a877496169f01a13f
Do not call list_l3_agent_hosting_router if not supported

For the admin router view, list_l3_agent_hosting_router is called.
If this extension is not supported, it returns "resource not found"
and an error popup.
This api should not be called in case the extension is not supported.

Closes-bug: #1708081
Change-Id: I38f7386689d95dde3fc76bbf4d6b5e781b515878