Updated openstack/openstack Project: openstack/nova  ce53afce8c7605bef412a393f193414c448e92c8
Fix lists in process doc

A lot of our rst was actually misformatted with extra spaces indenting
the bullet lists. This wasn't very visually noticable in the
oslosphinx theme, but it's really bad in the new theme. There will
need to be lots of fixes here, but will do them one page at a time.

Part of bp: doc-migration

Change-Id: Ic1d1383148b005b884fdfa3e9a9658adcb13ee70