Updated openstack/openstack Project: openstack/nova  33617699108bb3eb9632af3800e918915a7e38c8
Fix list rendering in bdm doc.

What looked clear in the rst actually was far from clear when rendered
in HTML. The document was restructured a bit so all the options end up
in a single bullet list, and the combination description is a separate
section.

Part of bp: doc-migration

Change-Id: I2feee4018a332483658d24d299dbb04ec87f2df0