Updated openstack/openstack Project: openstack-infra/puppet-log_processor  662ae3777c9e96cc931e2f1847804732567a9905
Handle cases without a build_change field

Previously the mqtt topic generation always assumed a build_change was
present. However there are some cases where the isn't a build_change in
the metadata, like periodic, post, and release jobs. This commit handles
those edge cases so it uses the build queue in the topic instead of the
build_change. If that doesn't work the topic is just the project.

Change-Id: I26dba76e3475749d00a45b076d981778f885c339