Updated openstack/openstack Project: openstack/api-wg  e9be2965b289a07746521f2985144525dd7f3908
Update compatibility doc and references

The compatibility doc was out of date and making promises of having
additional information that is now provided by the interoperability
document. So we now link from compatibility to api_interoperability and
in discoverability where we had been linking to compatibility we now
link to the recent information on consuming the catalog.

Change-Id: I296d5b167ffad63a256a2b133ee9250710e2837b
Closes-Bug: #1593312