Updated openstack/openstack Project: openstack/keystone-specs  5006df160c93734970ebd532749e244c48bbf749
Bump application credentials to backlog

Due to time and resource constraints, coupled with ongoing
discussions regarding the intended behavior of application
credentials, we weren't able to land this during the Pike
release. Let's move this to the backlog, rehash our approach
during the PTG in Denver, and slate this for Queens.

Change-Id: I7ae97c5a5e51cefb2f0604cfe7883bb1b73cb99a