Updated openstack/openstack Project: openstack/tripleo-heat-templates  45c95100cfd3ae63ae86a67e468baca0c9758dab
Add environment for setting a custom domain name

Just setting CloudDomain won't make the domains used consistent.
There are a number of CloudName parameters that must be set as well.
This change adds a sample environment that includes all of those
parameters so it is easy to set everything consistently.

Also fixes the description of CloudNameCtlplane to reflect the
actual use for that parameter.

Change-Id: I56d1c1c5619f83c16c4e8350aa84fccc3d748425