Updated openstack/openstack Project: openstack-infra/puppet-log_processor  88e0d21347b5492dba8c6c619f1d53cdbc9cb6db
Only send mqtt events for processed files

We were previously sending events for every file we attempted to
process, not just those that were processed and also for every single
log line event. This effectively doubled the io performed by the
logstash workers which seemed to slow the whole pipeline down. Trim it
down to only recording events for log files that are processed which
should significantly trim down the total number of events.

Change-Id: I0daf3eb2e2b3240e3efa4f2c7bac57de99505df0