Updated openstack/openstack Project: openstack-infra/project-config  aaf9dc5ef2c739506bc01ccdd53205cd9a50cae2
Add 1.0 feature branch in Gerritbot for ARA

There will be heavy and backwards incompatible work going on as
we work towards version 1.0 of ARA and it will be done in the 1.0
branch.
We intend to keep the master branch for the "stable" 0.14 release
which might end up getting bugfixes and dot releases in the meantime.

Change-Id: Ib6e6e3ed4f25af7ed64a212b6605ee42fdd5dfa3