Updated openstack/openstack Project: openstack-infra/project-config  8ad71685932bb148de875a2502e917a1a58a095a
Update puppet jobs to 4.8

There have been some recent merging of changes to things like stdlib
that require 4.6+. In tripleo we actually deploy with 4.8.  This change
updates the Puppet OpenStack and puppet-tripleo jobs to use 4.8 for unit
tests.

It should be noted that 4.5 equates to Puppet Enterprise 2016.2[0]
which has already reached limited support as of October of 2016[1].
4.8[2] is in PE 2016.5 which is newer than the current LTS version.

[0] https://docs.puppet.com/puppet/4.5/release_notes.html
[1] https://puppet.com/misc/puppet-enterprise-lifecycle
[2] https://docs.puppet.com/puppet/4.8/release_notes.html

Change-Id: I6091aacd692129789b95c45c0cde55e4a55ccb51