Updated openstack/openstack Project: openstack/tripleo-heat-templates  f96e07c909e6bc665ad816ab42a8e6fcd184c8aa
Adds environment file for ODL + SRIOV

Using the separate neutron-opendaylight and SRIOV env files do not work
because sriov includes using OVS agent (which ODL does not want or need)
and the default ODL env file has no Compute ML2 because it is not
needed.  Thus a new environment file is needed for deploying these 2
features in combination.

Closes-Bug: 1696667

Change-Id: I6f7a9368aa521de928c269619278c30acda03799
Signed-off-by: Tim Rozet <trozet@redhat.com>