Updated openstack/openstack Project: openstack/tripleo-heat-templates  3899491b1cb9fd39f4c2df00c8b955bd6040dc04
Change the directory for haproxy certs/keys to be service-specific

This moves the directories containing the certs/keys for haproxy one step
further inside the hierarchy. This way we will be able to bind-mount
this certificate into the container without bind-mounting any other
certs/keys from other services.

bp tls-via-certmonger-containers

Depends-On: Iba3adb9464a755e67c6f87d1233b3affa8be565a
Change-Id: I73df8d442b361cb5ef4e343b4ea2a198a5b95da9