Updated openstack/openstack Project: openstack/nova  2e75b9102662afd5ba40040f56f2d13061b78303
Fix 409 handling in report client when deleting inventory

Change I30b51a3235e69f4a2993052f7f9ef853858f2d65 regressed
a check added in I10b22606f704abcb970939fb2cd77f026d4d6322
which is meant to avoid dumping errors in the logs when
the scheduler report client fails to delete compute node
resource provider inventory since some resource class inventory
is still in use. It's all very fragile since the check
in the report client is dependent on a regex for the
error message coming back from the Placement API, but until
we have error codes in the API responses, this is what we
have to deal with.

This change puts the original error message back in the
Placement API and adds notes for awareness.

Change-Id: I9a9af0e3e04c7ea7314b7077a681d6a111de6e47
Closes-Bug: #1708031