Updated openstack/openstack Project: openstack/cinder  0ae42ac801beb88e5c570ca044c3128088a3094c
Refresh az cache if target az is not found in cache

When creating volume, Cinder use az cache to validate az. This
will lead a problem: If a new cinder-volume node with a different
az is added to the OpenStack environment or admin change an exist
cinder-volume's az, Creating volume with this az will fail until
the cache is refreshed.

This patch will refresh the cached az if the previous creating
task failed to find the target.

Change-Id: I3e884af1499ea2ddea1a3603d3d09c31a1f62811
Cloese-bug: #1693084