Updated openstack/openstack Project: openstack-infra/elastic-recheck  712e8626cb9e1699ed906638951008f8f8f52058
Remove query for bug 1497006

Remove this query not because it doesn't have any hits but because it
follows a symptom of underlying problems not the actual underlying
problem. In this case it is a failure of having ansible present which is
installed via pip into a virtualenv. We should have queries that track
virtualenv and pip failures instead.

We have recently reworked the pip installation failure tracking so this
should be quite robust now and there are no known virtualenv failures so
that should cover our bases.

Change-Id: Icd949cd101489646631326d98fdb9042a52a19aa