Updated openstack/openstack Project: openstack-infra/shade  697cf58c826c56ee1599d23ec1271933e8aae54b
Make QoS rules required parameters to be not optional

Some of QoS rules parameters, like max_kbps for bandwidth limit rule,
dscp_mark for dscp marking rule or min_kbps for minimum bandwidth
are not optional on Neutron's side so it shouldn't be optional in
shade client.
This commit changes them to be required in create methods.

Change-Id: I6919690d5fe024698dc76bf9808ec8ada80d2ae8