Updated openstack/openstack Project: openstack-infra/subunit2sql  b99350137b76a6f56a39a92cd464cec01f46f421
Switch to using stestr

stestr is a fork of testrepository, designed specifically to concentrate
on being a dedicated test runner. [1] The testrepository project is
basically not active anymore and has several long standing bugs and
useablility issues. stestr is still actively maintained and fixes a
large number of those issues with testrepository. The best part is that
stestr supports using subunit2sql as a repository backend.

[1] https://github.com/mtreinish/stestr

Change-Id: I39f5db8f77de027c18917215c761dca69d8cbea0