Updated openstack/openstack Project: openstack/cinder  a6ee8d4c7d7640ad5ca389969424024521c514e2
Mark Coho driver as unsupported

The Coho CI has not reported success in over 51 days and is past our
CI required period.

Checking name: Coho Data Storage CI
  first seen: 2017-05-28 13:30:07 (51 days, 11:35:18 old)
  last seen: 2017-07-03 19:46:18 (15 days, 5:19:07 old)
  last success:  None
  Job dsvm-tempest-coho-cinder-driver 0% success out of 240 comments S=0, F=240

Per Cinder's non-compliance policy, this patch marks the driver as
unsupported and deprecated and it will be removed in the next
cycle if the issue is not corrected.

Change-Id: Ic4e2ee817cb6e34c120760af3045b500b9d2a99f