Updated openstack/openstack Project: openstack-infra/devstack-gate  a1a40066042b0b6b3e59a53bf5ccf471caa8be08
Revert "Use caching reverse proxy for pypi"

This reverts commit 18cfa2598c0d5a04cb6543610b6a0c4cd5cd4bde.

This didn't catastrophically fail but we are seeing that when new
releases to python packages are made our cached indexes for these
packages are lagging in noticeable ways. We can likely get away with
this for the most part but it desynchonizes the afs volume releases
across regions and could possibly cause problems with timing as we get
into release time.

Change-Id: Ib50545f9cab0ba583127450e2c83a0c464c6f68a