Updated openstack/openstack Project: openstack/horizon  09563b6b2baa58e1337a5be71451bcce2c962959
Update the documentation link for doc migration

This commit updates links affected by doc migration.
Note that other links are not touched in this commit.
Fixing other broken links or changing http to https
should be covered by a separate patch.

Also convert http links to docs.openstack.org to https per reviewer
request (though it is not directly related to doc-migration).

Note that this commit does not touch files under releasenotes/notes/.
If we touch release notes in older releases, reno will considers
it as a release note in the current release.
(Pike release notes are clean up in a separate patch.)

Change-Id: Iba9bd465ef08014a4972584cf4da6e7d77961119