Updated openstack/openstack Project: openstack/cinder  2a27fe404a71837b194dee328253aad6cbbf506c
VMAX driver - seamless upgrade from SMI-S to REST

In the Pike release, the VMAX driver uses Unisphere RESTAPI to
communicate with the backend storage array, instead of SMI-S which
was used in previous releases. This has resulted in some architecture
and naming changes, which currently do not allow for seamless upgrade
for existing snapshots and 'in-use' volumes from SMI-S based driver
to REST based driver (e.g. upgrading from Ocata to Pike). This patch
rectifies the issue and supports attaching/detaching of upgraded
volumes, allows slo/workload to be set in the xml file, and supports
legacy snapshots.

Change-Id: I9064bc836da6fa8b8e73dbbeff811cc738de360a
Closes-Bug: #1703341