Updated openstack/openstack Project: openstack/horizon  50354f024841a27fda078128bd021788c6dc69c2
Add some notices on handling release notes

During the recent work on release notes I got some useful lessons on reno.
It is worth documented. (Note that they will be added to the reno docs
but it is nice to have it in horizon doc too)

Also adds a link to the release note section in the contributor guide
to the release note page so that developers can find information on
the release notes more easily.

Change-Id: I0c5bd5f5e70d6c1be4eb8f8e895312205177774e