Updated openstack/openstack Project: openstack-infra/project-config  7e57a4799e523e4ef5ed0563c6fe663e0304fbbf
Run TripleO scenario002 for Telemetry services

We added Telemetry services in Scenario002:
https://review.openstack.org/#/c/474969

So now we're able to test Telemetry when using Swift as a storage
backend for Gnocchi (scenario001 uses RBD).

This patch runs tripleo-scenario002 where scenario001 was already
running and it also adds the files where we want to run the jobs.

Change-Id: Id13a9ee27c9444c0d698345947da3e34ab877881