Updated openstack/openstack Project: openstack-infra/project-config  85117b31ef3e2bece4b830f08f2a45e8425edbd1
Skip tempest job of Mogan if not related to code

For Mogan gate, tempest job currently run for patches which is not
related code. That wastes gate resources and developers need to
wait for much time even if they propose doc patches.
This patch adds all-files-match-any like the other projects.

Change-Id: I82261c708a9163150d5d56cf179ad2d79b9d896c