Updated openstack/openstack Project: openstack-infra/elastic-recheck  1c6fcbe0ae72c69f39e55d657db7aff479228822
Remove query for bug 1697808

This bug was due to unreliable DNS which was caused by using ipv4 dns
resolvers in osic which had to go through NAT which occasionally failed.
We no longer have access to any osic resources and this bug was fixed
when we did. Therefore just remove the query.

Change-Id: I26e8cfc5fcb970bfa1dc214718e1796c3bf38b02