Updated openstack/openstack Project: openstack/nova  b93432e907a7cd33cae8f81fcaee06f5d0743166
Add Contributor Guide section page

In order to wrangle together a top level TOC that makes sense, stop
including everything in the TOC (ideally we could convince sphinx to
not warn on that).

This creates a Contributor Documentation landing page with deep links
to all the relevant documents. It removes all those links from the
front page, and the TOC. They can be found by deep links or search,
and it is a less confusing initial experience for folks.

Part of bp: doc-migration

Change-Id: If5e7940ddd0ae3316f7475742c02abfe3df28ac4