Updated openstack/openstack Project: openstack/sahara  6581e6570e48cab8aa317735fa60220797323f99
Restructure the documentation according the new spec

Group the existing pages in the new top-level categories; each
category can be seen as a separate document and it has its own
index file.
The content of the pages was not changed, with the obvious
exception of the links between pages.

The autogenerated configuration has not been added yet to
configuration/; it will be fixed in a future commit.

At the same time, as suggested by the doc team, consistently
use only one separator in file names (dash, '-') instead of a mix
of dashes, dots and underscores. This may break even more links on
the Internet, but we are breaking them anyway by moving files.
Redirects can be set, but not in this commit.

Closes-Bug: #1706184
Change-Id: I5a10378d9da2603d617ad4193ea8d90e2afc5104