Updated openstack/openstack Project: openstack/cinder  e0c91299b9cf160939b79f04e8dbe88c35e15bc7
Make Cinder's Landing Page look like Keystone's

The index.rst page for Cinder was not well suited
to the new openstackdocstheme that was created.  This
resulted in the page just looking like a disorganized
set of links with no visually defined topics or
explanation.

With this patch I emulate the look of the Keystone page
by properly indicating the different topics in the page
and add some text to describe each of the sections.  I also
update some of the old 'Block Storage' pages to use a 'Cinder'
heading for consistency.

Change-Id: I341e8340b960946a86f4ff8fbee3624531b375af