Updated openstack/openstack Project: openstack/nova  f742baa4c3bd54995237387e5f4ef0710977ebe6
Use wsgi-intercept in OSAPIFixture

To lessen the amount of eventlet socket handling present in the
functional tests, run osapi_compute via wsgi-intercept rather than
spawning an eventlet server to run it in a "thread".

To get this to work three main changes are made to the fixture:

* gain access to the wsgi app via Loader().load_app. This is used
  because it involves the least amount of manipulation of CONF. The
  init_application in nova/api/openstack/wsgi_app.py tries to
  re-read configuration.

* manually register the osapi_compute service (previously built
  in as a result of using WSGIService)

* Override the TCPKeepAliveAdapter keystone session adapter. It's
  use of socket_options does not work will with the way that six and
  wsgi-intercept mess with the HTTPConnection class.

The change in integrated_helpers is required because the new fixture
must provide access to the WSGI app in a different way.

Change-Id: Ia91fd7d7605dfb1a6f4376b1680cd26ea410d0f3
Related-Bug: #1705753